backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
numpy
xgboost
yfinance
statsmodels
scikit-learn
google-generativeai
langchain-experimental
langchain-google-genai
langgraph
selenium
tqdm
schedule
sqlalchemy
psycopg2-binary
alembic